simple compilation & simulation scripts added (sv2v -> iverilog)

+++ script/compile.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-riscv64-unknown-elf-as -march=rv32ic $1 -o rom.o
-riscv64-unknown-elf-objcopy -O binary -j .text rom.o ../rom/rom.bin
-rm rom.o
+if [ -z "$1" ]
+then
+    echo "compile.sh needs an assembly source!"
+else
+    riscv64-unknown-elf-as -march=rv32ic $1 -o rom.o
+    riscv64-unknown-elf-objcopy -O binary -j .text rom.o ../rom/rom.bin
+    rm rom.o
+fi
